backend/requirements.txt
# GlobalEvents backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
alembic>=1.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis>=5.0.0
python-jose[cryptography]
passlib[bcrypt]
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
